// Zidio_Frontend/store/types.ts
export type UpdateAction = (index: number, key: string, value: any) => void;

export interface Store<ObjectType> {
  data: Array<ObjectType>;
  setData?: (list: Array<ObjectType>) => void;
  add: (obj: ObjectType) => void;
  update: UpdateAction;
}